refactor: JSON 강의 데이터 적재 모듈 패키지 정리
